chore: refactor for trpc
- replace nextjs `app` directory with `pages`
- move all files to `src` directory
- update all imports
