--- app/page.tsx
@@ -1,9 +1,0 @@
-import Image from "next/image";
-
-export default function Home() {
-  return (
-    <div>
-      <h1 className="text-2xl">Homepage</h1>
-    </div>
-  );
-}
