--- components/ui/list-item.tsx
@@ -1,32 +1,0 @@
-import { Copy } from "lucide-react";
-import { ComponentPropsWithoutRef, ElementRef, forwardRef } from "react";
-
-import { cn } from "@/lib/utils";
-import { NavigationMenuLink } from "@radix-ui/react-navigation-menu";
-
-export const ListItem = forwardRef<
-  ElementRef<"a">,
-  ComponentPropsWithoutRef<"a">
->(({ className, title, children, ...props }, ref) => {
-  return (
-    <>
-      <NavigationMenuLink asChild>
-        <a
-          ref={ref}
-          className={cn(
-            "block select-none space-y-1 rounded-md p-3 leading-none no-underline outline-none transition-colors hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground",
-            className
-          )}
-          {...props}
-        >
-          <div className="text-sm font-medium leading-none">{title}</div>
-          <p className="text-sm leading-snug line-clamp-2 text-muted-foreground">
-            {children}
-          </p>
-        </a>
-      </NavigationMenuLink>
-    </>
-  );
-});
-
-ListItem.displayName = "ListItem";
