--- lib/utils/trpc.ts
@@ -1,45 +1,0 @@
-import { getFetch, httpBatchLink } from "@trpc/client";
-import { createTRPCNext } from "@trpc/next";
-import superjson from "superjson";
-import { AppRouter } from "@/lib/server/routers/app.routes";
-
-export const trpc = createTRPCNext<AppRouter>({
-  config({ ctx }) {
-    const url = process.env.NEXT_PUBLIC_VERCEL_URL
-      ? `https://${process.env.NEXT_PUBLIC_VERCEL_URL}`
-      : "http://localhost:3000/api/trpc/";
-
-    return {
-      queryClientConfig: {
-        defaultOptions: {
-          queries: {
-            staleTime: 5 * 1000,
-          },
-        },
-      },
-      headers() {
-        if (ctx?.req) {
-          return {
-            ...ctx.req.headers,
-            "x-ssr": "1",
-          };
-        }
-        return {};
-      },
-      links: [
-        httpBatchLink({
-          url,
-          fetch: async (input, init?) => {
-            const fetch = getFetch();
-            return fetch(input, {
-              ...init,
-              credentials: "include",
-            });
-          },
-        }),
-      ],
-      transformer: superjson,
-    };
-  },
-  ssr: true,
-});
