+++ src/pages/api/[trpc].ts
@@ -1,0 +1,9 @@
+import * as trpcNext from "@trpc/server/adapters/next";
+
+import { appRouter } from "@/server/routers/app.routes";
+import { createContext } from "@/server/createContext";
+
+export default trpcNext.createNextApiHandler({
+  router: appRouter,
+  createContext,
+});
