+++ src/pages/index.tsx
@@ -1,0 +1,11 @@
+import { NextPageWithLayout } from "./_app";
+
+const IndexPage: NextPageWithLayout = () => {
+  return (
+    <div className="grid w-full h-full grid-cols-1 gap-4">
+      <h1 className="text-2xl text-center capitalize">IndexPage</h1>
+    </div>
+  );
+};
+
+export default IndexPage;
