+++ src/server/controllers/user.controller.ts
@@ -1,0 +1,19 @@
+import type { Context } from "@/server/createContext";
+import { TRPCError } from "@trpc/server";
+
+export const getMeHandler = ({ ctx }: { ctx: Context }) => {
+  try {
+    const user = ctx.user;
+    return {
+      status: "success",
+      data: {
+        user,
+      },
+    };
+  } catch (err: any) {
+    throw new TRPCError({
+      code: "INTERNAL_SERVER_ERROR",
+      message: err.message,
+    });
+  }
+};
