+++ src/server/services/user.service.ts
@@ -1,0 +1,58 @@
+import { Prisma, User } from "@prisma/client";
+
+import customConfig from "@/server/config/default";
+import { prisma } from "@/server/utils/prisma";
+import redisClient from "@/server/utils/connectRedis";
+import { signJwt } from "@/server/utils/jwt";
+
+export const createUser = async (input: Prisma.UserCreateInput) => {
+  return (await prisma.user.create({
+    data: input,
+  })) as User;
+};
+
+export const findUser = async (
+  where: Partial<Prisma.UserCreateInput>,
+  select?: Prisma.UserSelect
+) => {
+  return (await prisma.user.findFirst({
+    where,
+    select,
+  })) as User;
+};
+
+export const findUniqueUser = async (
+  where: Prisma.UserWhereUniqueInput,
+  select?: Prisma.UserSelect
+) => {
+  return (await prisma.user.findUnique({
+    where,
+    select,
+  })) as User;
+};
+
+export const updateUser = async (
+  where: Partial<Prisma.UserWhereUniqueInput>,
+  data: Prisma.UserUpdateInput,
+  select?: Prisma.UserSelect
+) => {
+  return (await prisma.user.update({ where, data, select })) as User;
+};
+
+export const signTokens = async (user: Prisma.UserCreateInput) => {
+  // 1. Create Session
+  redisClient.set(`${user.id}`, JSON.stringify(user), {
+    EX: customConfig.redisCacheExpiresIn * 60,
+  });
+
+  // 2. Create Access and Refresh tokens
+  const access_token = signJwt({ sub: user.id }, "accessTokenPrivateKey", {
+    expiresIn: `${customConfig.accessTokenExpiresIn}m`,
+  });
+
+  const refresh_token = signJwt({ sub: user.id }, "refreshTokenPrivateKey", {
+    expiresIn: `${customConfig.refreshTokenExpiresIn}m`,
+  });
+
+  return { access_token, refresh_token };
+};
