+++ src/server/utils/jwt.ts
@@ -1,0 +1,30 @@
+import jwt, { SignOptions } from "jsonwebtoken";
+
+import customConfig from "@/server/config/default";
+
+export const signJwt = (
+  payload: Object,
+  key: "accessTokenPrivateKey" | "refreshTokenPrivateKey",
+  options: SignOptions = {}
+) => {
+  const privateKey = Buffer.from(customConfig[key], "base64").toString("ascii");
+  return jwt.sign(payload, privateKey, {
+    ...(options && options),
+    algorithm: "RS256",
+  });
+};
+
+export const verifyJwt = <T>(
+  token: string,
+  key: "accessTokenPublicKey" | "refreshTokenPublicKey"
+): T | null => {
+  try {
+    const publicKey = Buffer.from(customConfig[key], "base64").toString(
+      "ascii"
+    );
+    return jwt.verify(token, publicKey) as T;
+  } catch (error) {
+    console.log(error);
+    return null;
+  }
+};
